safari-534.53-branch/LayoutTests/fast/canvas/webgl/gl-getstring.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
  "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
<title>WebGL gl.getParameter Strings Conformance Tests</title>
<link rel="stylesheet" href="../../js/resources/js-test-style.css"/>
<script src="resources/desktop-gl-constants.js" type="text/javascript"></script>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<canvas id="canvas" width="2" height="2"> </canvas>
<script>
description("This test checks getParameter returns strings in the correct format");

debug("");
debug("Canvas.getContext");

var gl = create3DContext(document.getElementById("canvas"));
if (!gl) {
  testFailed("context does not exist");
} else {
  testPassed("context exists");

  debug("");
  checkPrefix("WebGL 1.0", "VERSION");
  checkPrefix("WebGL GLSL ES 1.0", "SHADING_LANGUAGE_VERSION");
  shouldBeNonNull("gl.getParameter(gl.VENDOR)");
  shouldBeNonNull("gl.getParameter(gl.RENDERER)");
  shouldBe("gl.getError()", "gl.NO_ERROR");
}

function checkPrefix(expected, enum_val) {
  var s = gl.getParameter(gl[enum_val]);
  if (s != null &&
      s.length >= expected.length &&
      s.substring(0, expected.length) == expected) {
    testPassed("getParameter(gl." + enum_val + ") correctly started with " + expected);
  } else {
    testFailed("getParameter(gl." + enum_val + ") did not start with " + expected);
  }
}

debug("");
successfullyParsed = true;

</script>
<script src="../../js/resources/js-test-post.js"></script>

<script>
</script>

</body>
</html>
